[wip] separate plots that use 'group by' into a new file
